Implement FunctionExpression's.

+++ ctjs/ctjs/ast/position.cc
@@ -1,5 +1,5 @@
 #include "ctjs/ast/position.h"
 
 namespace ctjs::ast {
-Position::Position(int line, int column) : line_(line), column_(column) {}
+Position::Position(size_t line, size_t column) : line_(line), column_(column) {}
 }  // namespace ctjs::ast
